Cleaned up and simplified code
